Added comments, checks for the memory file, unstuck bouncing, and reverse function
I made it so that if the robot is ever backed into a corner it will just go back the way it came from to avoid crashing, though it will continue that way and might backtrack a lot, its advisable we figure out away to pick a non-diagonal direction, to continue in.

I made the program check to see if memorySpace.mat exists before calling on it to grab pD from.

I added a function called reverse it takes in direction value and gives the direction value corresponding to the opposite of the keypad.

The robot will get stuck if it approaches a lone black diagonally because the corner of its FOV is invisible, so we need a way of checking inside the 3x3 "safety square" if our approach is still going to be diagonally.

I think I might call this short term straight line bot instead, cause its not really making a beeline towards anything.