Update and rename firstcode.m to ShortTermStraightLine.m
The name change is kind of to mark it as a contender to be a working base for the code that we submit next week. Also in the future, if you want to work on this piece of code as your base, branch then duplicate the file and rename it as your own.