Create advanced_roomba.m
Authors: Zack Weinstein, Kiley Stahl
Theory: Pick a random direction to start in. Only analyze the 3x3 space around the bot, and only pay attention to block in the area the bot will be on the next step. If there is a block there pick a direction on the opposite side of the keypad that isn't the direction the bot is coming from. If there is a block there then it will pick a direction on the opposite side from that block.