Created LongDistanceRoomba, and updated Readme to include it
Theory: Break sections of FOV into groups (NE, N, NW, W, etc). Only pay attention to the group immediately in front of the bot on the next step. If there is a wall in any of these spaces (which can be determined using the sum command) mark this as impermeable and check a one of the 2 adjacent directions to the opposite of the current direction. Keep checking until you find a permeable direction.

I believe this is our best script so far.