added in robot 3, updated to new verison of robot 2, slams
Slams keep track of how far in a given cardinal direction the robot should go. For example left starts 5 (for now) which is the left side of the optibox. If we ever find a wall thats in an x-position greater than five (or the current left(y)) at some y-level we change left(y) to be the x-value of the wall. There is slam for the bottom, left, right, and top in that order.